feat(aep-api): export public secretsprovider port for overlay
