+++ services/aep-api/internal/clients/secretmanagersvc/types.go
@@ -16,72 +16,37 @@
 
 package secretmanagersvc
 
-import "errors"
+import "github.com/wso2/aep/aep-api/secretsprovider"
 
 // ErrSecretNotFound is returned when a secret does not exist.
-var ErrSecretNotFound = errors.New("secret not found")
+var ErrSecretNotFound = secretsprovider.ErrSecretNotFound
 
 // ErrNotManaged is returned when attempting to delete a secret not managed by this client.
-var ErrNotManaged = errors.New("secret not managed by this client")
+var ErrNotManaged = secretsprovider.ErrNotManaged
 
 // ErrMetadataNotFound is returned when secret metadata does not exist.
-var ErrMetadataNotFound = errors.New("secret metadata not found")
+var ErrMetadataNotFound = secretsprovider.ErrMetadataNotFound
 
 // ErrNotSupported is returned when an operation is not supported by the provider.
-var ErrNotSupported = errors.New("operation not supported by this provider")
+var ErrNotSupported = secretsprovider.ErrNotSupported
 
 // SecretMetadata contains metadata for a secret.
-type SecretMetadata struct {
-	// ManagedBy identifies who manages this secret.
-	// Used to prevent accidental deletion of secrets created outside this client.
-	ManagedBy string `json:"managedBy,omitempty"`
-
-	// Labels are optional key-value pairs for additional metadata.
-	Labels map[string]string `json:"labels,omitempty"`
-}
+type SecretMetadata = secretsprovider.SecretMetadata
 
 // SecretInfo contains information about a secret without the actual values.
-type SecretInfo struct {
-	// ID is the unique identifier for the secret (e.g., secretReferenceName).
-	ID string `json:"id"`
-
-	// Name is the logical name of the secret.
-	Name string `json:"name,omitempty"`
-
-	// Keys is the list of keys available in the secret (without values).
-	Keys []string `json:"keys,omitempty"`
-
-	// Labels are optional key-value pairs for additional metadata.
-	Labels map[string]string `json:"labels,omitempty"`
-
-	// CreatedAt is the timestamp when the secret was created.
-	CreatedAt string `json:"createdAt,omitempty"`
-}
+type SecretInfo = secretsprovider.SecretInfo
 
 // StoreConfig holds configuration for secret store backends.
-type StoreConfig struct {
-	// Provider is the name of the provider to use (e.g., "openbao", "vault", "aws").
-	Provider string `json:"provider"`
-
-	// OpenBao contains OpenBao/Vault-specific configuration.
-	OpenBao *OpenBaoConfig `json:"openbao,omitempty"`
-}
+type StoreConfig = secretsprovider.StoreConfig
 
 // OpenBaoConfig contains configuration for OpenBao/Vault.
-// Only KV v2 secrets engine is supported.
-type OpenBaoConfig struct {
-	// Server is the OpenBao server address (e.g., "https://openbao.example.com").
-	Server string `json:"server"`
-
-	// Path is the mount path for the KV secrets engine (e.g., "secret").
-	Path string `json:"path"`
-
-	// Auth contains authentication configuration.
-	Auth *OpenBaoAuth `json:"auth"`
-}
+type OpenBaoConfig = secretsprovider.OpenBaoConfig
 
 // OpenBaoAuth contains authentication configuration for OpenBao.
-type OpenBaoAuth struct {
-	// Token is a static token for authentication.
-	Token string `json:"token,omitempty"`
-}
+type OpenBaoAuth = secretsprovider.OpenBaoAuth
+
+// SecretLocation identifies where a secret lives in the KV hierarchy.
+type SecretLocation = secretsprovider.SecretLocation
+
+// ParseKVPath inverts KVPath for the six legal location shapes.
+var ParseKVPath = secretsprovider.ParseKVPath
