+++ services/aep-api/secretsprovider/doc.go
@@ -1,0 +1,26 @@
+// Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com).
+//
+// WSO2 LLC. licenses this file to you under the Apache License,
+// Version 2.0 (the "License"); you may not use this file except
+// in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+// Package secretsprovider is the public secrets-delivery port for aep-api.
+//
+// Overlay modules and OSS main implement Provider and supply
+// Options.SecretsProvider without importing internal/ clients or platform
+// packages. The high-level SecretManagementClient in internal/clients/
+// secretmanagersvc consumes this port and consults SecretReferenceManager
+// when authoring OpenChoreo SecretReferences.
+//
+// This package must not import secretmanagersvc or app (no import cycle).
+package secretsprovider
