+++ services/aep-api/secretsprovider/errors.go
@@ -1,0 +1,31 @@
+// Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com).
+//
+// WSO2 LLC. licenses this file to you under the Apache License,
+// Version 2.0 (the "License"); you may not use this file except
+// in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package secretsprovider
+
+import "errors"
+
+// ErrSecretNotFound is returned when a secret does not exist.
+var ErrSecretNotFound = errors.New("secret not found")
+
+// ErrNotManaged is returned when attempting to delete a secret not managed by this client.
+var ErrNotManaged = errors.New("secret not managed by this client")
+
+// ErrMetadataNotFound is returned when secret metadata does not exist.
+var ErrMetadataNotFound = errors.New("secret metadata not found")
+
+// ErrNotSupported is returned when an operation is not supported by the provider.
+var ErrNotSupported = errors.New("operation not supported by this provider")
